Change DataId from an Int to a Byte

+++ RED/Interfaces/ISubscribe.cs
@@ -2,6 +2,6 @@
 {
     public interface ISubscribe
     {
-        void Receive(int dataId, byte[] data);
+        void Receive(byte dataId, byte[] data);
     }
 }
